requirements.txt
asgiref==3.4.1
Django==2.2.20
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
